pyproject.toml
[project]
name = "python-gd-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.12.4",
    "google-api-python-client>=2.184.0",
    "google-auth>=2.41.1",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp>=1.16.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
